Correct enter of steps area
Former-commit-id: e6bdc02a8198b33030c348ce96e9e561ee0db337

+++ PokeHunt/Classes/Settings/Settings.swift
@@ -81,7 +81,7 @@ class Settings : NSObject, NSCoding {
         }
     }
 
-    var stepsArea: UInt8 = 30 {
+    var stepsArea: Int = 30 {
         didSet{
             save()
             notify()
@@ -137,6 +137,6 @@ class Settings : NSObject, NSCoding {
         self.showPokemons = aDecoder.decodeBoolForKey(kShowPokemonsKey)
         self.showPokestop = aDecoder.decodeBoolForKey(kShowPokestopKey)
         self.showJym = aDecoder.decodeBoolForKey(kShowJymKey)
-        self.stepsArea = UInt8(aDecoder.decodeIntForKey(kStepsAreaKey))
+        self.stepsArea = Int(aDecoder.decodeIntForKey(kStepsAreaKey))
     }
 }
